Fixed the map bug on Giveaway page

+++ src/main/java/se/kth/sda/skeleton/post/Post.java
@@ -48,7 +48,7 @@ public class Post {
 
     @Column
     @ElementCollection
-    private List<Integer> position;
+    private List<Double> position;
 
 
     @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
@@ -162,11 +162,11 @@ public class Post {
         this.location = location;
     }
 
-    public List<Integer> getPosition() {
+    public List<Double> getPosition() {
         return position;
     }
 
-    public void setPosition(List<Integer> position) {
+    public void setPosition(List<Double> position) {
         this.position = position;
     }
 }
